Vendor files properly
